refactor: fix code issues
