--- src/modules/ItineraryMakerModule/constants/index.ts
@@ -1,12 +1,0 @@
-export const SECTION_COLORS = [
-  { class: 'bg-red-700', hex: '#b91c1c' },
-  { class: 'bg-green-700', hex: '#15803d' },
-  { class: 'bg-blue-700', hex: '#1d4ed8' },
-  { class: 'bg-yellow-700', hex: '#a16207' },
-  { class: 'bg-purple-700', hex: '#7e22ce' },
-  { class: 'bg-pink-700', hex: '#be185d' },
-  { class: 'bg-orange-700', hex: '#c2410c' },
-  { class: 'bg-teal-700', hex: '#0f766e' },
-  { class: 'bg-indigo-700', hex: '#4338ca' },
-  { class: 'bg-emerald-700', hex: '#047857' },
-]
